move backgroundServer.js
